Add OpenGL backend

+++ imj-base/test/Test/Imj/Ease.hs
@@ -11,10 +11,10 @@ testEase = do
   putStrLn ""
   test quartInOut
 
-test :: (Float -> Float) -> IO ()
+test :: (Double -> Double) -> IO ()
 test ease = mapM_ (\v -> putStrLn $ show v) $ map ease $ map (\i -> fromIntegral i / 10.0) [(0 :: Int)..10]
 
-quartInOut :: Float -> Float
+quartInOut :: Double -> Double
 quartInOut time =
     if time < 0.5
     then        1 / 2 *  2**4 * time * time  * time  * time
